refactor(tree): refactor the fuzz test code (#14842)
